Rearrange expressions test files and directories to clearly split between visitor, parsing and linear expressions tests
